requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
torch
nltk
accelerate
PyPDF2
tqdm
